Finished Q-Learning (Fix reward?)

+++ learning/src/learning.cpp
@@ -765,7 +765,7 @@ void getStateFromIndex(int index){
     int num_elems = discr_level + 1;
     robot_state.distance_d = ((int)index / (int)(pow(num_elems,2) * pow(2,2)))-1;
     robot_state.angle_d = ((index % (int)(pow(num_elems,2) * pow(2,2))) / (int)(num_elems * pow(2,2)))-1;
-    robot_state.height_d = ((index % (int)(pow(num_elems,2) * pow(2,2)) % (int)(num_elems * pow(2,2))) / pow(2,2))-1;
+    robot_state.height_d = ((index % (int)(pow(num_elems,2) * pow(2,2)) % (int)(num_elems * pow(2,2))) / (int)pow(2,2))-1;
     robot_state.object_picked = (index % (int)(pow(num_elems,2) * pow(2,2)) % (int)(num_elems * pow(2,2)) % (int)pow(2,2)) / 2;
     robot_state.folded = (index % (int)(pow(num_elems,2) * pow(2,2)) % (int)(num_elems * pow(2,2)) % (int)pow(2,2) % 2);
 }
